requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-core==2.3.2
google-cloud-datastore==2.15.1
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httptools==0.5.0
idna==3.4
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
